Fix the 64-bit scriptable_handle.h.
Review URL: http://codereview.chromium.org/271018

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@842 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

+++ src/trusted/plugin/srpc/scriptable_handle.h
@@ -284,7 +284,7 @@ class ScriptableHandle: public ScriptableHandleBase {
     dprintf(("ScriptableHandle::RemoveProperty(%p, %s)\n",
              static_cast<void*>(obj),
              PortablePluginInterface::IdentToString(
-                 reinterpret_cast<int>(name))));
+                 reinterpret_cast<uintptr_t>(name))));
 
     PortablePluginInterface* intf = unknown_handle->plugin_interface_;
     if (NULL == intf->nacl_instance()) {
